Chores: - Add Linux environment check for daemonset command - Write platform specific functions for using ns package
