+++ cmd/internal/kube/daemonset/daemonset.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"os"
+	"runtime"
 	"runtime/debug"
 	"sync"
 	"time"
@@ -74,6 +75,11 @@ type Daemonset struct {
 }
 
 func StartDaemonset() error {
+	// Check if the agent is running in a linux environment
+	if runtime.GOOS != "linux" {
+		return errors.New("This command is only supported on linux images")
+	}
+
 	// Initialize the front client
 	postmanInsightsVerificationToken := os.Getenv("POSTMAN_INSIGHTS_VERIFICATION_TOKEN")
 	frontClient := rest.NewFrontClient(
